CanInterface::set_can_params switch to use InterfaceCanParams
Instead of introducing a new struct SetCanParams for use with the new
set_can_params method switch to use the already existing
InterfaceCanParams struct. So there is one struct less to maintain.
Reusing the previously existing InterfaceCanParams struct has the
downside, that it contains some members, that can not be set but only
queried. These members need to be an Option now, so that they can be
None for the case we use InterfaceCanParams for setting params.